backward: fp64 products in the d_theta reduction -- fp32 lost 8/10 seeds
The d_theta sum is ILL-CONDITIONED: with random data the result is ~512 while the sum of
absolute terms is ~200000, a condition number near 400, so fp32 product rounding is amplified
~400x. Measured on K=4 all-fp32 over 10 seeds, the kernel was LESS accurate than eager in 8 of
them -- systematic, not tail-statistic luck (d_theta is a scalar, so it has no tail).

Products now formed in fp64 before the tile reduction; partials were already fp64.

Note this is a different situation from the forward, where the kernel wins on MEAN error 15/15
seeds and only loses on MAX 5/15 -- there the max is a tail statistic over 262k elements at the
fp32 rounding floor and the coin-flip is expected. d_theta has no such excuse.

Speed cost not yet measured.

+++ kernels/sm75/residual_add.py
@@ -186,7 +186,7 @@ def _res_add_bwd(
         # product STORED in bf16 before the reduction runs. Summing the fp32 product
         # instead left 0.5% relative error on the scalar gradient.
         tl.store(PART + pid * spart + 0,
-                 tl.sum(tl.sum(go * s, axis=1), axis=0).to(tl.float64))
+                 tl.sum(tl.sum(go.to(tl.float64) * s.to(tl.float64), axis=1), axis=0))
         if NEED0:
             tl.store(DS0 + offs_t[:, None] * d0_t + offs_h[None, :],
                      (c * go).to(DS0.dtype.element_ty), mask=mask)
@@ -201,7 +201,7 @@ def _res_add_bwd(
         # product STORED in bf16 before the reduction runs. Summing the fp32 product
         # instead left 0.5% relative error on the scalar gradient.
         tl.store(PART + pid * spart + 1,
-                 tl.sum(tl.sum(go * s, axis=1), axis=0).to(tl.float64))
+                 tl.sum(tl.sum(go.to(tl.float64) * s.to(tl.float64), axis=1), axis=0))
         if NEED1:
             tl.store(DS1 + offs_t[:, None] * d1_t + offs_h[None, :],
                      (c * go).to(DS1.dtype.element_ty), mask=mask)
@@ -216,7 +216,7 @@ def _res_add_bwd(
         # product STORED in bf16 before the reduction runs. Summing the fp32 product
         # instead left 0.5% relative error on the scalar gradient.
         tl.store(PART + pid * spart + 2,
-                 tl.sum(tl.sum(go * s, axis=1), axis=0).to(tl.float64))
+                 tl.sum(tl.sum(go.to(tl.float64) * s.to(tl.float64), axis=1), axis=0))
         if NEED2:
             tl.store(DS2 + offs_t[:, None] * d2_t + offs_h[None, :],
                      (c * go).to(DS2.dtype.element_ty), mask=mask)
@@ -231,7 +231,7 @@ def _res_add_bwd(
         # product STORED in bf16 before the reduction runs. Summing the fp32 product
         # instead left 0.5% relative error on the scalar gradient.
         tl.store(PART + pid * spart + 3,
-                 tl.sum(tl.sum(go * s, axis=1), axis=0).to(tl.float64))
+                 tl.sum(tl.sum(go.to(tl.float64) * s.to(tl.float64), axis=1), axis=0))
         if NEED3:
             tl.store(DS3 + offs_t[:, None] * d3_t + offs_h[None, :],
                      (c * go).to(DS3.dtype.element_ty), mask=mask)
